Docked new widget on right side of IDA for scroll
Figured out through testing how to dock a widget on the right side of
the IDA window and set its relative size to the other widgets. Created a
complete copy of MAGIC_form as MAGIC_sync_scroll. The newly redocked
widget is simply just a second version of the first widget in a
different spot, it ill be modified in the future